[QuyenLe] edit database, random data for db, FUNCTION_PROCEDURE, code KiemHang

+++ trunk/PawnShopManager/PawnShopManager/Entity/LoaiHang.cs
@@ -4,7 +4,7 @@ namespace PawnShopManager.Entity
 {
    class LoaiHang
    {
-      public Int16 loaihang_Id { get; set; }
+      public Int32 loaihang_Id { get; set; }
       public String loaihang_Ten { get; set; }
    }
 }
