+++ trunk/PawnShopManager/PawnShopManager/Entity/ThemTien.cs
@@ -7,8 +7,8 @@ namespace PawnShopManager.Entity
       public Int32 them_Id { get; set; }
       public Int32 hd_Id { get; set; }
       public String them_NgayTao { get; set; }
-      public float them_TienLai { get; set; }
-      public float them_Tien { get; set; }
-      public Int16 them_HinhThuc_TraLai { get; set; }
+      public Double them_TienLai { get; set; }
+      public Double them_Tien { get; set; }
+      public Int32 them_HinhThuc_TraLai { get; set; }
    }
 }
