+++ trunk/PawnShopManager/PawnShopManager/Entity/ThuChi.cs
@@ -5,9 +5,9 @@ namespace PawnShopManager.Entity
    class ThuChi
    {
       public Int32 tc_Id { get; set; }
-      public Int16 tc_Loai { get; set; }
+      public Int32 tc_Loai { get; set; }
       public String tc_NgayTao { get; set; }
-      public float tc_Tien { get; set; }
+      public Double tc_Tien { get; set; }
       public String tc_Note { get; set; }
    }
 }
